requirements.txt
amqp==2.5.1
attrs==19.1.0
billiard==3.6.1.0
celery==4.3.0
certifi==2019.9.11
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.2.5
django-auth-kerberos==1.2.5
django-celery-results==1.1.2
django-cors-headers==3.1.0
django-js-asset==1.2.2
django-mptt==0.10.0
djangorestframework==3.10.3
djangorestframework-simplejwt==4.3.0
docutils==0.15.2
et-xmlfile==1.0.1
get==2019.4.13
graphviz==0.13
idna==2.8
importlib-metadata==0.22
inflection==0.3.1
itypes==1.1.0
jdcal==1.4.1
Jinja2==2.10.1
jsonschema==3.0.2
kombu==4.6.4
ldap3==2.6.1
Markdown==3.1.1
MarkupSafe==1.1.1
more-itertools==7.2.0
numpy==1.17.2
openapi==1.1.0
openapi-codec==1.3.2
openpyxl==2.6.3
pandas==0.25.1
ply==3.11
post==2019.4.13
psycopg2==2.8.3
psycopg2-binary==2.8.3
public==2019.4.13
pyasn1==0.4.7
PyJWT==1.7.1
pykerberos==1.2.1
pyrsistent==0.15.4
python-dateutil==2.8.0
pytz==2019.2
query-string==2019.4.13
request==2019.4.13
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.25.3
vine==1.3.0
zipp==0.6.0
